[apex] Detect Schema namespace in TypeShadowsBuiltInNamespace
- Rename from TypeShadowsSystemNamespace to TypeShadowsBuiltInNamespace
- Consider additionally classes from Schema namespace
- Make visitor static
